Where I left off
This is the point where I last left off. I can't remember what I was
doing 3 months ago at 3AM, so not much of a description to give. It
appears that in the time since I last worked on this, Atmosphere has
made several updates to the source, so this EEL will likely be heavily
modified.